--- src/features/universities/containers/UniversityContainer.tsx
@@ -1,8 +1,0 @@
-import React from 'react';
-import FindUniversity from '../components/FindUniversity';
-
-const UniversityContainer = () => {
-  return <FindUniversity />;
-};
-
-export default UniversityContainer;
