--- src/features/universities/containers/UniversityInfoModalContainer.tsx
@@ -1,8 +1,0 @@
-import React from 'react';
-import UniversityInfoModal from '../components/FindUniversity/modals/InfoModal';
-
-const UniversityInfoModalContainer = () => {
-  return <UniversityInfoModal name="연세대학교 체육교육학과" subTitle="수능 일반 전형" competition="14.05:1" />;
-};
-
-export default UniversityInfoModalContainer;
