+++ src/features/universities/containers/UniversityInformationModalContainer.tsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import UniversityInformationModal from '../components/FindUniversity/modals/InformationModal';
+
+const UniversityInfoModalContainer = () => {
+  return <UniversityInformationModal name="연세대학교 체육교육학과" subTitle="수능 일반 전형" competition="14.05:1" />;
+};
+
+export default UniversityInfoModalContainer;
