+++ src/features/universities/containers/UniversityTableContainer.tsx
@@ -1,0 +1,14 @@
+import React from 'react';
+import UniversityTable from '../components/FindUniversity/UniversityTable';
+import UniversityInfoModalContainer from './UniversityInformationModalContainer';
+
+const UniversityTableContainer = () => {
+  return (
+    <>
+      <UniversityTable />
+      <UniversityInfoModalContainer />
+    </>
+  );
+};
+
+export default UniversityTableContainer;
